// frontend/src/app/openapi/model/history-event.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */

import { Project } from './project';
import { User } from './user';
import { EventType } from './event-type';


export interface HistoryEvent { 
    user: User;
    executor: User | null;
    project: Project | null;
    execution_time: string;
    event_type: EventType;
    reason: string | null;
    id: number;
}
export namespace HistoryEvent {
}